+++ lib/web_autostart_settings_method_channel.dart
@@ -1,11 +1,9 @@
-import 'dart:io';
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
 import 'autostart_settings_platform_interface.dart';
 
-/// An implementation of [AutostartSettingsPlatform] that uses method channels.
+/// Not implemented for web
 class MethodChannelAutostartSettings extends AutostartSettingsPlatform {
   /// The method channel used to interact with the native platform.
   @visibleForTesting
@@ -13,16 +11,10 @@ class MethodChannelAutostartSettings extends AutostartSettingsPlatform {
 
   @override
   Future<bool> canOpen({required bool autoStart, required bool batterySafer}) async {
-    if (!Platform.isAndroid) {
-      return false;
-    }
-    return await methodChannel.invokeMethod<bool>('canOpen', {'autoStart': autoStart, 'batterySafer': batterySafer}) ?? false;
+    return false;
   }
   @override
   Future<bool> open({required bool autoStart, required bool batterySafer}) async {
-    if (!Platform.isAndroid) {
-      return false;
-    }
-    return await methodChannel.invokeMethod<bool>('open', {'autoStart': autoStart, 'batterySafer': batterySafer}) ?? false;
+    return false;
   }
 }